+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.9.3] - 2025-07-17
+### Changed
+- remove numbers at start of frame astm
+
+### Fixed
+- last line of HL7 message for Lab29 transactions
+- convertRSP_K11toASTM
+
 ## [0.9.2] - 2025-07-10
 ### Added
 - type_cnx : socket_E1381